move components into separate files
