+++ react/components/DatePicker/DatePicker.tsx
@@ -1,97 +1,9 @@
 import { useState } from "react";
-import {
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  CalendarDaysIcon,
-} from "@heroicons/react/16/solid";
+import { CalendarDaysIcon } from "@heroicons/react/16/solid";
 
-import Week from "./Week";
-import { getWeeksForMonth, MONTH_NAMES, WEEKDAY_NAMES } from "../../lib/dates";
+import { getWeeksForMonth } from "../../lib/dates";
 import { format } from "date-fns";
-
-interface HeaderProps {
-  month: number;
-  year: number;
-  handleMonthDecrease: (month: number) => void;
-  handleMonthIncrease: (month: number) => void;
-}
-
-const Header = ({
-  month,
-  year,
-  handleMonthDecrease,
-  handleMonthIncrease,
-}: HeaderProps) => (
-  <div className="flex justify-between w-full">
-    <ChevronLeftIcon
-      className="w-6 h-6 cursor-pointer hover:bg-neutral-300 rounded-md transition-[background-color]"
-      onClick={() => handleMonthDecrease(month - 1)}
-    />
-    <div className="select-none">
-      {MONTH_NAMES[month]} {year}
-    </div>
-    <ChevronRightIcon
-      className="w-6 h-6 cursor-pointer hover:bg-neutral-300 rounded-md transition-[background-color]"
-      onClick={() => handleMonthIncrease(month + 1)}
-    />
-  </div>
-);
-
-interface MonthProps {
-  month: number;
-  year: number;
-  handleMonthDecrease: (month: number) => void;
-  handleMonthIncrease: (month: number) => void;
-  weeksForMonth: (Date | null)[][];
-  selectedDate: Date;
-  setSelectedDate: (date: Date) => void;
-  setIsDatePickerOpen: (isOpen: boolean) => void;
-}
-
-const Month = ({
-  month,
-  year,
-  handleMonthDecrease,
-  handleMonthIncrease,
-  weeksForMonth,
-  selectedDate,
-  setSelectedDate,
-  setIsDatePickerOpen,
-}: MonthProps) => {
-  return (
-    <div className="flex justify-center items-center w-full flex-col gap-2 border border-gray-300 rounded-md shadow-sm p-2 mt-2">
-      <div className="flex flex-col w-full">
-        <Header
-          month={month}
-          year={year}
-          handleMonthDecrease={handleMonthDecrease}
-          handleMonthIncrease={handleMonthIncrease}
-        />
-        <div className="flex select-none">
-          {WEEKDAY_NAMES.map((day, index) => (
-            <div
-              key={index}
-              className="flex justify-center flex-1 text-neutral-500"
-            >
-              {day}
-            </div>
-          ))}
-        </div>
-        {weeksForMonth.map((week, index) => (
-          <Week
-            key={index}
-            days={week}
-            selectedDate={selectedDate}
-            setSelectedDate={(date) => {
-              setSelectedDate(date);
-              setIsDatePickerOpen(false);
-            }}
-          />
-        ))}
-      </div>
-    </div>
-  );
-};
+import Month from "./Month";
 
 interface DatePickerProps {
   month?: number;
